(PC-15055)[PRO] refactor: move getStepsOffer to useOfferFormSteps
