+++ pro/src/core/Offers/hooks/index.ts
@@ -1,1 +1,2 @@
 export { default as useQuerySearchFilters } from './useQuerySearchFilters'
+export { default as useOfferFormSteps } from './useOfferFormSteps'
