+++ pro/src/core/Offers/index.ts
@@ -1,2 +1,3 @@
 export * from './constants'
+export * from './hooks'
 export * from './utils'
