+++ pro/src/screens/OfferIndividual/Informations/Informations.tsx
@@ -3,6 +3,7 @@ import React from 'react'
 import { useHistory } from 'react-router-dom'
 
 import Spinner from 'components/layout/Spinner'
+import { useOfferFormSteps, OFFER_FORM_STEP_IDS } from 'core/Offers'
 import { TOffererName } from 'core/Offerers/types'
 import { TOfferIndividualVenue } from 'core/Venue/types'
 import { OfferFormLayout } from 'new_components/OfferFormLayout'
@@ -13,8 +14,8 @@ import {
 } from 'new_components/OfferIndividualForm'
 
 import { ActionBar } from '../ActionBar'
-import { OFFER_FORM_STEP_IDS, fakeOffer } from '../constants'
-import { getStepsOffer } from '../utils/steps'
+import { fakeOffer } from '../constants'
+
 
 export interface IInformationsProps {
   createOfferAdapter: (
@@ -36,7 +37,7 @@ const Informations = ({
   const history = useHistory()
 
   // call getStep with offer when this screen get it as prop
-  const { activeSteps } = getStepsOffer(fakeOffer)
+  const { activeSteps } = useOfferFormSteps(fakeOffer)
   const handleNextStep = async () => formik.handleSubmit()
 
   const onSubmit = async (formValues: IOfferIndividualFormValues) => {
