+++ pro/src/screens/OfferIndividual/constants.ts
@@ -1,10 +1,3 @@
-export enum OFFER_FORM_STEP_IDS {
-  INFORMATIONS = 'informations',
-  STOCKS = 'stocks',
-  SUMMARY = 'recapitulatif',
-  CONFIRMATION = 'confirmation',
-}
-
 export const fakeOffer = {
   id: 'AL4Q',
   status: 'ACTIVE',
